+++ Packages/com.nickmaltbie.openkcc/OpenKCC/Character/KCCMovementEngine.cs
@@ -33,7 +33,7 @@ namespace nickmaltbie.OpenKCC.Character
     /// </summary>
     [RequireComponent(typeof(Rigidbody))]
     [RequireComponent(typeof(IColliderCast))]
-    public class KCCMovementEngine : MonoBehaviour, IKCCConfig
+    public class KCCMovementEngine : MonoBehaviour, IKCCConfig, ISerializationCallbackReceiver
     {
         /// <summary>
         /// Default value for distance from ground player is considered grounded.
@@ -74,6 +74,18 @@ namespace nickmaltbie.OpenKCC.Character
         /// Max launch speed of player form moving ground.
         /// </summary>
         public const float DefaultMaxLaunchVelocity = 5.0f;
+
+        /// <summary>
+        /// Serialization version for the KCCMovementEngine.
+        /// </summary>
+        public const string CurrentSerializationVersion = "v1.0.0";
+
+        /// <summary>
+        /// Current serialization version of this serialized KCCMovementEngine.
+        /// </summary>
+        [HideInInspector]
+        [SerializeField]
+        internal string serializationVersion;
 
         /// <summary>
         /// Height of a step that the player can climb up.
@@ -128,6 +140,11 @@ namespace nickmaltbie.OpenKCC.Character
         /// </summary>
         public float MaxPushSpeed => 100.0f;
 
+        /// <summary>
+        /// Layermask for computing player collisions.
+        /// </summary>
+        public LayerMask layerMask = IColliderCast.DefaultLayerMask;
+
         /// <inheritdoc/>
         public bool CanSnapUp => GroundedState.OnGround;
 
@@ -162,6 +179,9 @@ namespace nickmaltbie.OpenKCC.Character
         /// Current grounded state of the character.
         /// </summary>
         public KCCGroundedState GroundedState { get; protected set; }
+
+        /// <inheritdoc/>
+        public LayerMask LayerMask => layerMask;
 
         /// <summary>
         /// Collider cast for player shape.
@@ -202,7 +222,8 @@ namespace nickmaltbie.OpenKCC.Character
                 transform.rotation,
                 -Up,
                 SnapDown,
-                ColliderCast);
+                ColliderCast,
+                layerMask);
             transform.position += Vector3.ClampMagnitude(delta, MaxSnapDownSpeed * unityService.fixedDeltaTime);
         }
 
@@ -291,7 +312,9 @@ namespace nickmaltbie.OpenKCC.Character
             transform.position += ColliderCast.PushOutOverlapping(
                 transform.position,
                 transform.rotation,
-                MaxPushSpeed * unityService.fixedDeltaTime);
+                MaxPushSpeed * unityService.fixedDeltaTime,
+                layerMask,
+                QueryTriggerInteraction.Ignore);
 
             // Allow player to move
             KCCBounce[] bounces = moves.SelectMany(move =>
@@ -373,7 +396,8 @@ namespace nickmaltbie.OpenKCC.Character
                     transform.rotation,
                     -Up,
                     SnapDown,
-                    ColliderCast);
+                    ColliderCast,
+                    layerMask);
 
                 groundCheckPos += snapDelta;
             }
@@ -383,7 +407,8 @@ namespace nickmaltbie.OpenKCC.Character
                 transform.rotation,
                 -Up,
                 GroundCheckDistance,
-                out IRaycastHit hit);
+                out IRaycastHit hit,
+                layerMask);
 
             Vector3 normal = hit.normal;
 
@@ -403,6 +428,26 @@ namespace nickmaltbie.OpenKCC.Character
                 maxWalkAngle: MaxWalkAngle);
 
             return GroundedState;
+        }
+
+        public void OnBeforeSerialize()
+        {
+            // Nothing needed here.
+        }
+
+        public void OnAfterDeserialize()
+        {
+            // If the serialization version is unset, update it.
+            if (string.IsNullOrEmpty(serializationVersion))
+            {
+                serializationVersion = CurrentSerializationVersion;
+
+                // Set default value for layer mask
+                if (layerMask == default)
+                {
+                    layerMask = ~0;
+                }
+            }
         }
     }
 }